requirements.txt
python-telegram-bot[all]==21.9
python-dotenv==1.1.0
playwright==1.49.1
mcp[cli]>=1.26.0
requests
gTTS>=2.5.0
chromadb>=0.4.0
dateparser>=1.2.0
feedparser>=6.0.0
tweepy>=4.14.0
praw>=7.7.0
